+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Changed
+- `JSONSerializer`: added serialization of `Pair` and `Triple`
+- `JSONSerializerTest`: corresponding tests
 
 ## [0.4] - 2019-05-14
 ### Changed