+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - `JSONSerializerTest`: corresponding tests
 - `JSONDeserializer`: added deserialization of `Pair` and `Triple`
 - `JSONDeserializerTest`: corresponding tests
+- `JSONDeserializer`: added checking of types marked nullable
+- `JSONDeserializerTest`: corresponding tests
 
 ## [0.4] - 2019-05-14
 ### Changed